Fixed compilation on Windows due to symbol export issues

Shared windows builds work, but static does not.

* remotes/origin/dev/fix-windows-build:
  Added to 0.1.2 release notes file
  added MSVC multiprocessing build flag
  Added ocv module exports
  Added exports to vxl module
  Added exports to proj module
  Add -Wall to gnu flags, fix warning with algo def template spacialization
  fix core algorithm exporting
  added to error message to make use of exception message
  fixed conversion warning
  Added export macros to relevent structures in CORE
  Default to shared libs ON on windows
  Fixed flags injection string
  Added position independent code flag when building static
  Add flags structures to CMake
  Fix symbol definition fixing windows build

+++ maptk/core/algo/track_features.cxx
@@ -147,7 +147,7 @@ simple_track_features
   std::vector<unsigned> unmatched;
   std::back_insert_iterator<std::vector<unsigned> > unmatched_insert_itr(unmatched);
   std::set_difference(boost::counting_iterator<unsigned>(0),
-                      boost::counting_iterator<unsigned>(vf.size()),
+                      boost::counting_iterator<unsigned>(static_cast<unsigned int>(vf.size())),
                       matched.begin(), matched.end(),
                       unmatched_insert_itr);
   BOOST_FOREACH(unsigned i, unmatched)
